requirements.txt


MetaTrader5>=5.0.4682
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
colorlog>=6.0.0
